Replace Saxophone with Acoustic Guitar prototype
Deleted saxophone file and replaced it with guitar acoustic. The instrument is functional, as in it plays pluck synth, but the output does not sound like a guitar. The UI is half way done, the notes to be played does appear along with guitar image, but the grid and finger placement needs to be added.